Extracting the date pickers from the MessageListFragment and moving them to ScheduledMessageUtils
